move dirty form check to its own file
